Relocate validateLogin to user, Add href navigation on .pug files, Made changes to /login
