+++ auth.js
@@ -1,0 +1,9 @@
+const loginUser = (req, res, user) => {
+  req.session.auth = {
+    userId: user.id,
+  };
+};
+
+module.exports = {
+  loginUser,
+};
