+++ routes/users.js
@@ -2,7 +2,9 @@ var express = require('express');
 var router = express.Router();
 const db = require('../db/models');
 const bcrypt = require('bcryptjs');
-const { csrfProtection, asyncHandler, validateLogin, handleValidationErrors } = require('../utils');
+const { csrfProtection, asyncHandler, handleValidationErrors } = require('../utils');
+const { loginUser } = require('../auth');
+const { check, validationResult } = require('express-validator');
 
 /* GET users listing. */
 router.get('/', function(req, res, next) {
@@ -16,28 +18,46 @@ router.get('/login', csrfProtection, (req, res) => {
   });
 });
 
+const validateLogin = [
+  check("username")
+    .exists({ checkFalsy: true })
+    .withMessage("Please provide your login username."),
+  check("username")
+    .isLength({ max: 30 })
+    .withMessage("Your username cannot be longer than 30 characters."),
+  check("password")
+    .exists({ checkFalsy: true })
+    .withMessage("Please provide your email address."),
+]
+
 router.post('/login', csrfProtection, validateLogin, handleValidationErrors, asyncHandler(async (req, res, next) => {
   const {
     username,
     password
   } = req.body;
 
-  const user = db.User.findOne({
-    where: {
-      username
-    }
-  });
+  let errors = [];
 
-  if (user !== null) {
-    const passwordMatch = await bcrypt.compare(password, user.hashedPassword.toString());
-
-    if (passwordMatch) {
-      return res.redirect('/');
+  const validatorErrors = validationResult(req);
+  if (validatorErrors.isEmpty()) {
+    const user = await db.User.findOne({ where: { username } });
+    if (user !== null) {
+      const passwordMatch = await bcrypt.compare(password, user.hashedPassword.toString());
+      if (passwordMatch) {
+        loginUser(req, res, user);
+        return res.redirect('/');
+      }
     }
+    errors.push('Login failed for the provided username and password')
+  } else {
+    errors = validatorErrors.array().map((error) => error.msg)
   }
 
-  handleValidationErrors.push('Login failed for the combined email address and password');
-
+  res.render('login', {
+    title: 'Login',
+    csrfToken: req.csrfToken(),
+    errors,
+  });
 }));
 
 router.get('/signup', csrfProtection, (req, res) => {
@@ -62,12 +82,8 @@ router.post('/signup', csrfProtection, validateLogin, handleValidationErrors, as
 
   const hashedPassword = await bcrypt.hash(password, 10);
   user.hashedPassword = hashedPassword;
-  console.log(user)
   await user.save();
-  console.log(user)
-
   req.session.user = user;
-
   req.session.save(() => {
     res.redirect('/');
   })
