+++ views/index.pug
@@ -8,4 +8,7 @@ append head
 
 block content
   h1= title
-  p Welcome to your Project Skeleton
+  p Welcome to GoodLaughs
+  a(href="/users/login") Log In
+  div
+  a(href="/users/signup") Don't have an account?
